pyproject.toml
[tool.poetry]
name = "monorepo-factory"
version = "0.1.0"
description = ""
authors = ["Jacob Valdez <jacobfv123@gmail.com>"]
readme = "README.md"
packages = [{ include = "monorepo_factory" }]

[tool.poetry.dependencies]
python = "^3.10.9"
typer = "^0.9.0"
pyyaml = "^6.0.1"
rich = "^13.7.0"
toml = "^0.10.2"
python-box = "^7.1.1"
networkx = "^3.2.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
